Topページの作成

+++ db/migrate/20220601054327_create_momentum_counts.rb
@@ -2,7 +2,7 @@ class CreateMomentumCounts < ActiveRecord::Migration[6.0]
   def change
     create_table :momentum_counts do |t|
       t.string :work
-      t.decimal :magnification
+      t.float :magnification
 
       t.timestamps
     end
